+++ example/lib/main.dart
@@ -7,18 +7,19 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      //showSemanticsDebugger: true,
       title: 'Flutter Segment Display Example',
       theme: ThemeData(
         accentColor: Colors.red,
         brightness: Brightness.dark,
       ),
-      home: MyHomePage(title: 'Flutter Segment Display Example'),
+      home: const MyHomePage(title: 'Flutter Segment Display Example'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  const MyHomePage({Key key, this.title}) : super(key: key);
 
   final String title;
 
@@ -27,11 +28,28 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
+  int _counter;
+  SegmentStyle _segmentStyle;
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _counter = 0;
+    _segmentStyle = DefaultSegmentStyle(
+      enabledColor: Theme.of(context).accentColor,
+      disabledColor: Theme.of(context).accentColor.withOpacity(0.15),
+    );
+  }
 
   void _incrementCounter() {
     setState(() {
       _counter++;
+    });
+  }
+
+  void _changeSegmentStyle(SegmentStyle segmentStyle) {
+    setState(() {
+      _segmentStyle = segmentStyle;
     });
   }
 
@@ -41,25 +59,71 @@ class _MyHomePageState extends State<MyHomePage> {
       backgroundColor: Colors.black,
       appBar: AppBar(
         title: Text(widget.title),
+        actions: <Widget>[
+          PopupMenuButton<SegmentStyle>(
+            tooltip: "Segment style",
+            onSelected: _changeSegmentStyle,
+            itemBuilder: (BuildContext context) {
+              return [
+                PopupMenuItem<SegmentStyle>(
+                  value: DefaultSegmentStyle(
+                    enabledColor: Theme.of(context).accentColor,
+                    disabledColor:
+                        Theme.of(context).accentColor.withOpacity(0.15),
+                  ),
+                  child: const Text("Default"),
+                ),
+                PopupMenuItem<SegmentStyle>(
+                  value: RectSegmentStyle(
+                    enabledColor: Theme.of(context).accentColor,
+                    disabledColor:
+                        Theme.of(context).accentColor.withOpacity(0.15),
+                  ),
+                  child: const Text("Rect"),
+                ),
+                PopupMenuItem<SegmentStyle>(
+                  value: HexSegmentStyle(
+                    enabledColor: Theme.of(context).accentColor,
+                    disabledColor:
+                        Theme.of(context).accentColor.withOpacity(0.15),
+                  ),
+                  child: const Text("Hex"),
+                ),
+              ];
+            },
+          ),
+        ],
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
+            const Text('You have pushed the button this many times:'),
             Padding(
               padding: const EdgeInsets.all(8.0),
-              child: SixteenSegmentDisplay(
-                text: "$_counter",
-                textSize: 12.0,
-                backgroundColor: Colors.transparent,
-                segmentStyle: DefaultSegmentStyle(
-                  enabledColor: Theme.of(context).accentColor,
-                  disabledColor:
-                      Theme.of(context).accentColor.withOpacity(0.15),
-                ),
+              child: Column(
+                children: <Widget>[
+                  SevenSegmentDisplay(
+                    text: "$_counter",
+                    textSize: 12.0,
+                    backgroundColor: Colors.transparent,
+                    segmentStyle: _segmentStyle,
+                  ),
+                  const SizedBox(height: 10),
+                  FourteenSegmentDisplay(
+                    text: "$_counter",
+                    textSize: 12.0,
+                    backgroundColor: Colors.transparent,
+                    segmentStyle: _segmentStyle,
+                  ),
+                  const SizedBox(height: 10),
+                  SixteenSegmentDisplay(
+                    text: "$_counter",
+                    textSize: 12.0,
+                    backgroundColor: Colors.transparent,
+                    segmentStyle: _segmentStyle,
+                  ),
+                ],
               ),
             ),
           ],