+++ lib/src/segment_display_painter.dart
@@ -3,17 +3,17 @@ import 'package:segment_display/src/segment/segment.dart';
 
 /// [CustomPainter] used for painting display segments
 class SegmentDisplayPainter extends CustomPainter {
-  /// [Paint] style for enabled segments
+  /// [Paint] style for enabled segments.
   Paint _enabledPaint;
 
-  /// [Paint] style for disabled segments
+  /// [Paint] style for disabled segments.
   Paint _disabledPaint;
 
-  /// All segments that should be drawn
+  /// All segments that should be drawn.
   List<Segment> segments;
 
-  Path _enabledPath = Path();
-  Path _disabledPath = Path();
+  final _enabledPath = Path();
+  final _disabledPath = Path();
 
   /// Creates [SegmentDisplayPainter], sets paint styles and splits [segments]
   /// (paths) into enabled and disabled.
@@ -30,7 +30,7 @@ class SegmentDisplayPainter extends CustomPainter {
       ..style = PaintingStyle.fill
       ..color = disabledColor;
 
-    for (var segment in segments) {
+    for (final segment in segments) {
       if (segment.isEnabled) {
         _enabledPath.addPath(segment.path, Offset.zero);
       } else {