+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.0] - 2019-10-11
+* wrapped segment display with `Semantics` widget
+* updated example
+* added widget tests
+
 ## [0.2.0] - 2019-05-13
 * added **sixteen-segment** display
 * updated segment styles for sixteen-segment display